requirements.txt
fastapi
numpy
python-telegram-bot
uvicorn